+++ src/main/java/com/ingsoft/trueque/model/Persona.java
@@ -7,6 +7,15 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Entity
 @Table(name = "personas")
@@ -16,7 +25,7 @@ import lombok.experimental.SuperBuilder;
 @NoArgsConstructor
 @AllArgsConstructor
 @SuperBuilder
-public class Persona {
+public class Persona implements UserDetails {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -33,4 +42,36 @@ public class Persona {
     @Enumerated(EnumType.STRING)
     @NotNull
     private Rol rol;
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        if (getRol() == null) return Collections.emptyList();
+
+        List<GrantedAuthority> authorities = new ArrayList<>();
+
+        // Agregar el rol con el prefijo "ROLE_"
+        authorities.add(new SimpleGrantedAuthority("ROLE_" + getRol().name()));
+
+        // Agregar los permisos si existen
+        if (getRol().getPermisos() != null) {
+            authorities.addAll(
+                    getRol().getPermisos().stream()
+                            .map(permiso -> new SimpleGrantedAuthority(permiso.name()))
+                            .collect(Collectors.toList())
+            );
+        }
+
+        return authorities;
+    }
+
+
+    @Override
+    public String getPassword() {
+        return this.getClave();
+    }
+
+    @Override
+    public String getUsername() {
+        return this.getCorreo();
+    }
 }
